CARL ResourceManager: Move `ConnectError` into `resource::manager`.
