+++ opendut-carl/src/resource/persistence/storage/mod.rs
@@ -1,3 +1,4 @@
+use crate::resource::manager::error::ConnectError;
 use crate::resource::types::id::ResourceId;
 use crate::resource::types::resources::RelayedSubscriptionEvents;
 use crate::resource::types::Resource;
@@ -5,7 +6,7 @@ use crate::resource::persistence::error::PersistenceResult;
 use crate::resource::persistence::persistable::{Persistable, StorageKind};
 use crate::resource::persistence::{Db, Memory};
 use crate::resource::types::subscription::Subscribable;
-use crate::resource::{persistence, ConnectError};
+use crate::resource::persistence;
 use anyhow::anyhow;
 use prost::Message;
 use redb::backends::InMemoryBackend;
